# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Weather Demo Documentation",
        "description": "OpenApi description",
        "contact": {
            "email": "mogohax@users.noreply.github.com"
        },
        "license": {
            "name": "wtfpl",
            "url": "http://www.wtfpl.net/"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/weather": {
            "get": {
                "tags": [
                    "Weather"
                ],
                "summary": "Gets current weather data",
                "description": "Returns weather data for a given city",
                "operationId": "getWeather",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "City name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "api_key",
                        "in": "query",
                        "description": "The API key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "provider",
                        "in": "query",
                        "description": "Weather provider",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "openweather",
                                "random"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/WeatherDataResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "additionalProperties": {}
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CoordinatesResource": {
                "title": "CoordinatesResource",
                "description": "Coordiantes model",
                "properties": {
                    "lat": {
                        "description": "Latitude",
                        "type": "number"
                    },
                    "lon": {
                        "description": "Longitude",
                        "type": "number"
                    }
                },
                "type": "object"
            },
            "SpeedResource": {
                "title": "SpeedResource",
                "description": "Speed model",
                "properties": {
                    "value": {
                        "description": "Speed value",
                        "type": "number"
                    },
                    "units": {
                        "description": "Measuring units",
                        "type": "object",
                        "$ref": "#components/schemas/UnitsResource"
                    }
                },
                "type": "object"
            },
            "TemperatureResource": {
                "title": "TemperatureResource",
                "description": "Temperature model",
                "properties": {
                    "value": {
                        "description": "Temperature value",
                        "type": "number"
                    },
                    "units": {
                        "description": "Measuring units",
                        "type": "object",
                        "$ref": "#components/schemas/UnitsResource"
                    }
                },
                "type": "object"
            },
            "UnitsResource": {
                "title": "UnitsResource",
                "description": "Measuring units model",
                "properties": {
                    "name": {
                        "description": "Name of measuring unit",
                        "type": "string"
                    },
                    "symbol": {
                        "description": "Symbol of measuring unit",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "WeatherDataResource": {
                "title": "WeatherDataResource",
                "description": "Weather Data response model",
                "properties": {
                    "city": {
                        "description": "City data holder",
                        "properties": {
                            "name": {
                                "description": "Class WeatherDataResource",
                                "type": "string"
                            },
                            "location": {
                                "type": "object",
                                "$ref": "#components/schemas/CoordinatesResource"
                            }
                        },
                        "type": "object"
                    },
                    "temperature": {
                        "description": "Temperature data",
                        "properties": {
                            "current": {
                                "description": "Current temperature",
                                "type": "object",
                                "$ref": "#components/schemas/TemperatureResource"
                            },
                            "feels_like": {
                                "description": "Feels like temperature",
                                "type": "object",
                                "$ref": "#components/schemas/TemperatureResource"
                            }
                        },
                        "type": "object"
                    },
                    "wind_speed": {
                        "description": "Wind speed",
                        "type": "object",
                        "$ref": "#components/schemas/SpeedResource"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Weather",
            "description": "Weather actions"
        }
    ]
}